更新至食人魔

+++ token_change_team.py
@@ -44,8 +44,14 @@ def change_color_team(img_folder, save_folder, watermark=""):
             pass
         if os.path.exists(watermark):
             img = image_watermark(img, cv2.imread(watermark, -1))
-        cv2.imwrite(os.path.join(save_folder, img_name), img)
-        print(f"{img_name}阵营转换后图片处理完成")
+        if not os.path.exists(os.path.join(save_folder, img_name)):
+            cv2.imwrite(os.path.join(save_folder, img_name), img)
+            save_folder_new = os.path.join(save_folder.split("change_team")[0], "change_team_new")
+            os.makedirs(save_folder_new, exist_ok=True)
+            cv2.imwrite(os.path.join(save_folder_new, img_name), img)
+            print(f"{img_name}阵营转换后图片处理完成")
+        else:
+            print(f"{img_name}阵营转换后图片已存在")
 
 
 def change_color(img_origin_folder, save_path, watermark=""):
@@ -64,3 +70,5 @@ if __name__ == '__main__':
     save_path = r"./image"
     watermark = r"./image_all/watermark/Just_KeVin.png"
     change_color(img_origin_folder, save_path, watermark=watermark)
+
+    change_color(img_origin_folder, save_path)
